[BUGFIX] Fix new configuration options for fluid feature toggle
If the feature toggle fluidBasedPageModule is on, the functionality is different.
In that case, the new TSconfig options for setting DeepL / Google translate keys
are not active.

This change fixes the checks in Localization.js to make sure that the functionality
still works if the feature toggle fluidBasePageModule is on.

Making the configuration options available with the feature toggle can be handled
in a later patch.