[BUGFIX] Fix glossary sync persistance
Due to a bug in the DataHandlerHook, the glossary entries were not
synchronized to the table tx_wvdeepltranslate_domain_model_glossariessync.

This typically occurred in sites with several languages because a
for loop was exited too early.

Due to the glossary entries not being synchronized, the glossary was
not used for translation.

Resolves: #99