[TASK] Introduce GitHub actions to guard code work
This change introduces some GitHub action workflow
definitions as a firsts round-up. Due several issues
not all tests and php-versions are covered for now.

The defined workflow kicks on on pushes, pull-requests
and also adding scheduled runs (aka nightlies).

No functional tests are executed yet, as this needs
either a valid DeepL API key, and consuming limits
and contigents. In a dedicated step, it can be tried
to use the DeepL API MockServer for functional test.

See: https://github.com/DeepLcom/deepl-mock

**Note:** This is only a first start, after tackling
the issues and streamlining and house-keeping has been
done, this can be increased and further extended.