[TASK] Streamline tools configuration
This change moves tools configuration into directory
structure under `Build/`, separating each tool to it's
own subfolder.

On the one hand, this is a preparation moving forward
and aligning the CI tools and tests into known TYPO3
core structures. Additionally, it prepares for a later
revert to the core testing-framework along with the
core `runTests.sh` in parallel to the ddev logic.