[BUGFIX] Require PHP version in ext_emconf.php
EXT: wv_deepltranslate requires at least PHP 7.4, as stated in its
`composer.json` file. However, if the TYPO3 instance is a
non-Composer-based installation, the extension will be installed
regardless of the used PHP version, leading to hard crashes.

To catch this case, the PHP version constraint is added to
`ext_emconf.php`.

Fixes: #170