[TASK] Skip one test when `DEEPLMOCKAPISERVER` is used
The `DEEPLMOCKAPISERVER` is only a really rudimentary
api implementation and do not simulate "full" api in
any way. That means, that not all source/target languages
are supported like for production api.

Therefore, one test is skipped now when `DEEPLMOCKAPISERVER`
is in use because of not supporting `EN` as target language.
An additionally test is added to test from `EN` to `DE` in
anny case.

Furthermore, when mock server is used the supported content
and and translated strings are used for test expectations.