[TASK] Use `DEEPLMOCKAPISERVER` supported content and translation
The `DEEPLMOCKAPISERVER` is only a really rudimentary
api implementation and do not simulate "full" api in
any way. That means, that not all source/target languages
are supported like for production api.

Therefore, when mock server is used, the supported content
and and translated strings are used for test expectations.

[TASK] Donate fine grained method to retrieve api url parts

`\WebVision\WvDeepltranslate\Configuration` transports the
required configuration settings through the system, which
is a good approach.

This change donates more fine grained read methods to this
class, to access api url parts. This is a preparation for
an upcoming change.