[TASK] Respect the url scheme of configured api url
In a couple of places, `https` has been hardcoded enforced
as request scheme, ignoring the scheme of the configured
api url.

This may seem to be a "security" enforcement, but these
kind of overrides are not very user friendly and hidden
magic is always a pain in the ass for debugging issues.

This change now re-uses the configured api url scheme
when building api requests. One test is adopted to do
the same.

The donated methods on the `Configuration` class to
access api url parts are used for this.