[BUGFIX] Avoid doctrine/dbal `fetchAssociative()`
Due to the fact of the wide range of supported
TYPO3 core versions and the therefore opened
doctrine/dbal versions, the newer `fetchAssociative()`
methods should be avoided.

This change reverts to the longer existing `fetch()`
method to mitigate this issue. One `executeQuery()`
call is replaced with `execute()` for the same reason.

Resolves: #234