[TASK] Display correct versions in core switch script
For adding basic TYPO3 v12 core testing support
some commands have been copied. Sadly, not the
correct versions have been used for the output.
That means, it does not matches what is installed.

This change now fixes this to have the echo test
aligned to what is done. That avoids confusion.

Releases: main