[TASK] Remove obsolete version check from siteconfiguration tca
The siteconfiguration related tca override contained a
version switch to add the custom extension field only
for TYPO3 v11 and up.

Beside it should considered to drop that custom field,
the version check is removed because it is guaranteed
by the composer TYPO3 version constraint of the main
branch.

Releases: main