[BUGFIX] Guard `LocalizationUtility::translate` with `(string)` cast
The `LocalizationUtility::translate()` is used
in several places to translate something. The
method may return a string or null, where as
the used places usual expects to have a valid
string afterwards. For example, if the return
value is null and used to pass it to another
method like `htmlspecialchars()` this will
violates added argument types with PHP 8.0.

To avoid emitted PHP warnings call to this
method is guarded with a explicit `(string)`
type cast to mitigate this issue.

Releases: main, 3.0