[TASK] Avoid instantiating PageRenderer in `ext_localconf.php`
Registering custom backend javascript modules have been
done historically by instantiating the PageRenderer in
the `ext_localconf.php` for `TYPO3_MODE === 'BE'`. That
was a valid way to do so.

The `TYPO3_MODE` constant has been deprecated and will
be removed in TYPO3 v12, which means that adding the
JS module could not be restricted to the backend context.

Furthermore, instantiating the PageRenderer with TYPO3
v12 in the `ext_localconf.php` file would emit an error
exception due to the fact that the page renderer needs
a cache. Instantiating the CacheManager in an early stage
is not possible which further disqualifies to add the
JavaScript module in the `ext_localconf.php`.

Generally there are multiple ways to mitigate this issue,
but the decision has been made to use an `PageRenderer`
hook directly. That has the major benefit of getting the
PageRenderer passed and no need to instantiate or inject
it in any way. We get the application context deliverd
for free with it, so no need to directly acces the global
request object. That makes this the best win-win solution.

We do not need the additionally and later features of the
`\TYPO3\CMS\Core\Page\Event\BeforeJavaScriptsRenderingEvent`
and thus not using it.

This paves the way further for TYPO3 v12 compatibility.

**NOTE:** This changes does not validates if the JS module
          itself is compatible with TYPO3 v12. This should
          be done in a dedicated patch when general support
          is in a stage this can be checked.

Resolves: #222
Related: #128
Releases: main