[TASK] Replace deprecated `TYPO3_MODE` constant usage
The `TYPO3_MODE` constant defined is deprecated
and removed with TYPO3 v12. It's recommended to
use the `TYPO3` constant for the known file
`defined() or die()` guard.

This change replaces the simply guards with the
new constant, which is already available in the
TYPO3 v11 version.

**NOTE:**   One usage in `ext_localconf.php` is
            not touched as this requires a more
            enhanced replacement. That has been
            provided with #258 already.

Resolves: #259
Related: #128
Related: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66948
Releases: main