[BUGFIX] Ensure `docker compose v2 combat`
Executing unit and functional tests with
`docker compose 2.x` is rather slow.

That's a bug of docker compose itself,
albeit it's ot sure if they are taking
it as such. Background for this is the
changed parsing behaviour for service
commands, which now needs additional
escaping levels.

Linux supports to chop down commands
using the backslash and still taking
it as one command. For example:

```shell
SOME_ENV_VAR=1 \
  binary_to_execute <some_arguments>
```

Services defined for `runTests.sh` uses
this a lot. Because of the introduced
behaviour change, this is now longer
considered as one command and needs to
be additional escaped (the backslash).

Because of this, xdebug is always active
which slows down the test execution albeit
not explictly commanded.

docker compose v2 has a long standing
history of being absolutly unreliable
due to the fact randomly introducing
breaking things in minor and patch
level releases.

As docker-compose v1 has been announced
to be unsupported in the nearer time
we need to address this.

This change removes the usage of `\`
command chops in the docker-compose
`Build/testing-docker/docker-compose.yaml`
file.

Additionally, we set the env var `ORIGINAL_ROOT`
properly bootstrap unit- and functional tests.

Resolves: #257
Releases: main