[TASK] Migrate documentation rendering to `runTests.sh`
Having multiple execution points are not really
user friendly. Distributing config and docker
compose files is also not a really good idea.

This change now integrates the documentation
rendering container as `runTests.sh` task and
into the underlying `docker-compose.yaml`.

Furthermore, the `develop` tag is used as this
is the real latest rendering container provided
by the TYPO3 documentation team. This may change
at any point and can than be changed back. With
this, rendering warnings are solved due a bug
in the published `latest` container.

Thus removing the file in the project root
folder to clean that up a bit.

Resolves: #256
Releases: main