[TASK] Add a test to cover undefined fluid variable as argument
This change adds a basic test to show that null-coalsing operator
guard along with a string cast is needed when accessing viewhelper
arguments. Even if registered as required and of type string.

Note: Expected to have a failing functional test now !

The issue should be fixed in a dedicated follow up.