[TASK] Add UsageService and associated tests
Added a new UsageService to check the current usage and excessive
translation limit along with respective interface and associated tests.
It retrieves the API usage information from DeepL and checks if the
translation limit is exceeded. As a part of this, the service is
registered in the Services configuration, and the changes also include
an exception handling method in the Client class.