[TASK] Ensure correct code handling to avoid phpstan reportings
Recently new code has been added which triggered static
code analzer reportings. Sadly, the simple way of adding
them to the baseline has been choosen instead of handling
it correctly in the code.

`interfaces` are contracts by nature and naturally taken
as a surface for custom implementation and adjustments.
Therefore, it is eminent to address the purpose of an
interface when introduced and used to make clear if it
is a public or a internal contract.

This change modifies userland code to check properly before
accessing nullable properties. Interface contracts get now
proper contract descriptions.

Minor code cleanup and formatting as side-change.

Used command(s):

```terminal

> Build/Scripts/runTests.sh -p 7.4 -t 11 -s composerUpdate \
  && Build/Scripts/runTests.sh -p 7.4 -t 11 -s phpstanGenerateBaseline \
  && Build/Scripts/runTests.sh -p 8.1 -t 12 -s composerUpdate \
  && Build/Scripts/runTests.sh -p 8.1 -t 12 -s phpstanGenerateBaseline \
  && Build/Scripts/runTests.sh -p 8.1 -t 12 -s cgl
```