[TASK] Use the new php based documentation rendering container
The TYPO3 Documentation team hardly worked on implementing
a php based ReST file rendering container to increase the
audiance of possible maintainers and replace the good old
but aged sphinx based working horse. Great to know on this
is the fact that the base tool has been developed as a cross
open source project effort with people from the great symfony,
doctrine and other open source communities basing there docs
on good old ReST files.

The new container and render [1] also makes the usage a breeze
and is fully cross tool compatible (docker, podman,...) and
can be executed epherminal with only a single command:

```terminal
docker run --rm --pull always -v $(pwd):/project -it \
  ghcr.io/typo3-documentation/render-guides:latest --config=Documentation
```

Albeit still in the introduction phase, it is already used for
the one or other official guide and documentation and therefore
we should switch for the local review rendering of the extension
documentation towards the new container.

This change replaces the old container with the new one in
the still `docker-compose` based `Build/Scripts/runTests.sh`
to move forward to the modern approach and being a early
adopter on it.

The new renderer requires a new configuration format, but
gladly provides a migration for it, which has been used [2]:

```terminal
docker run --rm --pull always -v $(pwd):/project -it \
  ghcr.io/typo3-documentation/render-guides:latest \
  migrate Documentation
```

[1] https://github.com/TYPO3-Documentation/render-guides#usage-with-docker-via-supplied-container
[2] https://github.com/TYPO3-Documentation/render-guides/blob/main/Documentation/Migration/Index.rst#with-official-docker-container

Resolves: #56
Releases: main