[BUGFIX] Ensure higher precedence of partial ordering
In projects, loading sequence problems could occur
in the context of the page layout if the extension
`web-vision/enable-translated-content` was included.
As a result, the Deepl Translation Select field for
pages was not displayed in the page layout module.

This change uses a higher timestamp for the backend
layout override to ensure higher priority in file
matching to mitigate this override path ordering
issue.

Note: A corresponding remark to the cross extension
should be added to rise awareness not adding a higher
value then the base extension. Counts for all addon
extensions.