[BUGFIX] Ensure full compatible for proxy setting to retrieve Translator
With #331 the TYPO3 proxy setting has been validated
and set as `DeepL Guzzle Client` transport option to
have a proper configured proxy used.

The TYPO3 proxy setting however allows setting it as a
single string url or as array for different protocols,
which is not respected.

This change modifies the `AbstractClient->getTranslator()`
implementation to handle this double structure for named
option correctly now.

[1] https://docs.typo3.org/m/typo3/reference-coreapi/main/en-us/Configuration/Typo3ConfVars/HTTP.html#typo3ConfVars_http_proxy

Related: #331