[TASK] Refactor formality translation context handling.
The problem with global formality configuration is true.
That a target language has not been checked to see whether formality is supported.
An error response from the DeepL API is returned for unsupported languages,
which states ‘formality not supported’.

To enable better handling for formality, only supported target languages.
An additional formality field is added in the SiteConfiguration,
which reflects the options of the API.
The Formality field only appears for supported target languages,
which is checked against the supported languages API.

An upgrade wizard has been provided to simplify the conversion of the Formality configuration to the SiteConfiguration.

Additionally, a DTO TranslateContext has been introduced,
to centralise parameter handling for better handling of translation parameters.