[FEATURE] Better readable usage stats / silent return on lower severity
Refs #343

[FEATURE] Make translation limits more readable

[FEATURE] Follow up: More readable quota in toolbar

[FEATURE] Replace severity numbers with constants

[FEATURE] Enable color changes for DeepL action icon

Action icons dont have a fixed color set, in order to let TYPO3 set
the desired color - black in list actions, white in toolbar infos.

Change the DeepL SVG to not use black in all places.

[FEATURE] Determine severity based on quota usage

Calculate the message severity based on the API quota usage rate.
Matching popular quota warning levels if will notify above 90% usage,
and shown an error when the limit is exceeded.

Hide the flashmessage after a translation if the severity is low.

Refs #343

[FEATURE] Make widget more compact and use severity

Use severity levels in widget as well and reduce the height
to prevent unused whitespace.