[BUGFIX] Return buttons if page is not callable
The ButtonBarHook seems to be called on root (pid=0) or other possible
constellations, too. In these cases (and some other) the BackendUtility
can't return avalid page record and returns null instead. Following up
code didn't check, if a valid page record was loaded from the database,
which lead to errors. WIth this commit, a check on a `$page === null`
and hardening on possible not set array keys is added avoiding this issue