[BUGFIX] Working with glossary without a request not possible
The DeeplBackendUtility tried to detect the current working page by
guessing the page id via the global request. In some cases no request is
given, like on CLI calls. Avoiding errors not using the glossary when
not calling translation via backend, a new service
`ProcessingInstruction` is added, saving the current instructions inside
the global runtime cache. This information can now get detected in every
point of the translation and allows a correct detection of the current
working page including a valid glossary, if one exists.

A regression test is added demonstrating the glossary is now detected
and respected, even if no request is set up.