[TASK] Align TCA configuration for required column option
TYPO3 12.0 deprecated the old way to define a column
as required as part of the `eval` [1] and introduced
a new option as replacement [2].

Until now this has been handled within the corresponding
TCA Override file to keep support for TYPO3 v11 next to
TYPO3 v12.

This change removes the override and uses the new option
in the TCA file directly.

[1] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/12.0/Deprecation-97035-RequiredOptionInEvalKeyword.html
[2] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/12.0/Feature-97035-UtilizeRequiredDirectlyInTCAFieldConfiguration.html