[TASK] Introduce local patch `contribution` package/extension
To make contribution and workin on `EXT:deepltranslate_core` easier,
the provided ddev based local environment should be improved with
the ability to setup a basic working environment easily.

This change introduces a local development extension `EXT:contribution`
as preparation to give at least a simple and easy setup of a running
TYPO3 instance based on the TYPO3 `EXT:styleguide` core extension to
generate a backend and frontend tree / data. Combined with the TYPO3
v12 setup command this allows a simple instance generation.

For the start, the contribution extension only provide a cli command
to manage site root visibilities (on/off) having `EXT:styleguide`
generated trees in mind.