[TASK] Ensure to define all depending extension in functional tests
The next `typo3/testing-framework` includes additional
extension dependency check and sorting and will fail to
execute tests when not all extensions are configured as
`$testExtensionToLoad` or `$coreExtensionsToLoad` within
the functional test.

This change adds the missing extension configuration to
functional test cases to prepare the way for the next
`typo3/testing-framework` version and adding TYPO3 v13
suppport.