[TASK] Make ProcessingInstructions public service
As other services can use the DeeplService and translate by their own,
it's necessary to be able to override the ProcessingInstruction setting
DeepL mode to true. As this is not the default and inside the core extension
only needed for the command map, the service has to be made public, allowing
external services adjusting the instructions.