[TASK] Change Dropdown rendering for translations
Getting rid of x-classing is the goal and with v12 an event is provided
manipulating the header of the RecordListController.

In favour of using this event instead of x-classing the original class, an event
listener is added generating the dropdown and adding top-most of the recordlist.

This has effects on positioning of the translation dropdown, which now is
rendered before the core dropdown instead of next to it.