[TASK] Use simplified backend module template for `UsageWidget`
TYPO3 v12 introduced a simplified ext:backend ModuleTemplate API [1]
to avoid the use of `StandaloneView` in the backend context and it's
not a suprise that TYPO3 v13 continued on this path and deprecated
it in TYPO3 13.3 [2] finally removing it as breaking change in TYPO3
v14 [3].

Sounding quite challenging, it's quite easy to make the `UsageWidget`
for `EXT:dashboard` compatible with TYPO3 v12 and v13 using simplified
backend module template API directly with the template override [4]
feature already included.

The surrounding `<p>` tag in the fluid template for the usage text
is replaced with `<span>` to make the UX feeling more user-friendly.

[1] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/12.0/Feature-96730-SimplifiedExtbackendModuleTemplateAPI.html
[2] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/13.3/Deprecation-104773-CustomFluidViewsAndExtbase.html
[3] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/14.0/Breaking-105377-DeprecatedFunctionalityRemoved.html
[4] https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/12.0/Feature-96812-OverrideBackendTemplatesWithTSconfig.html