[TASK] Use custom DataHandler command instead of invalid cmdMap
Until now the DataHandler command map array has been misused
to transport custom `localize` command options relevant for
translation with DeepL, processed with a hook implementation
and using the original command.

TYPO3 v13 hardens the DataHandler even further, which is a
really good thing, and breaking fingers of developers mess
with the data monster - and loosing [1].

This change switch to a custom DataHandler command and using
the hook to handle the command directly by using the public
`DataHandler->localize()` method to mitigate the TYPO3 v13
exception thrown by `DataHandler->start()`.

Tests and link creation are aligned to use the new command
name and omit abitary and invalid options.

[1] https://review.typo3.org/c/Packages/TYPO3.CMS/+/83073